More

+++ src/polynomials/f2_poly_t.h
@@ -33,7 +33,7 @@ public:
 
   f2_poly_t(const std::string & s);
 
-  static f2_poly_t from_base_rep(unsigned b);
+  static f2_poly_t from_base_rep(uint64_t b);
   f2_poly_t prime_sfld_elt(int v) const;
   bool from_string(const char *string);
   int get_characteristic(void);
